% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.vdh.drug.overdose-package.R
\docType{package}
\name{dc.vdh.drug.overdose-package}
\alias{dc.vdh.drug.overdose}
\alias{dc.vdh.drug.overdose-package}
\title{dc.vdh.drug.overdose: Social Data Commons: Health, Behavior, Diet: Drug Overdose ED Visits}
\description{
Allows user to easily get drug overdose ED visit data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Zachary Alerte \email{rsu8zv@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
